% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm_cv.R
\name{blockSample}
\alias{blockSample}
\title{block sample of ppmData object}
\usage{
blockSample(object, p = 0.25, res = 1, seed = NULL)
}
\arguments{
\item{object}{ppmData A ppmData object}

\item{p}{numeric The thinning probability (Bernoulli variable); values are (0,1).}

\item{res}{numeric The resolution (size in map coordinates) of the blocks.}

\item{seed}{numeric The seed to set if you want to set.seed(). Default is NULL (no seed)}
}
\description{
Do a block sample of sites if you want a tougher k-folds test
}
\author{
Skipton Woolley
}
