% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm_predict.R
\name{createPredTiles}
\alias{createPredTiles}
\title{Function to generate tiles for large predictions using terra}
\usage{
createPredTiles(
  spatRasters,
  ntiles = 10,
  tileNames = "tile",
  tilesDir = "tiles"
)
}
\arguments{
\item{spatRasters}{A single or multiple layer rasters loaded using terra}

\item{ntiles}{The number of tiles to use for tiling in this will be squared so ntiles=10 equals 100 tiles in total.}

\item{tileNames}{The name to be given to each of the tiles e.g. "tile" would results in "tile1.tif" to "tileN.tif"}

\item{tilesDir}{The directory to save tiles}
}
\description{
Function to generate tiles for large predictions using terra
}
