% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm_cv.R
\name{cvTests}
\alias{cvTests}
\title{Assess performance of models using hold-out dataset}
\usage{
cvTests(object, slambda = NULL, ...)
}
\arguments{
\item{object}{A cvFit model object, which is essentially a bunch of model fitted to a holdout dataset.}

\item{slambda}{A numeric value to represent 'lambda.min', 'lambda.1se' or a set of lambda valies. Value(s) of the penalty parameter lambda at which predictions are required. Default is "lambda.min".}

\item{\dots}{Ignored}
}
\description{
This function produces summary performance measures for the glmnet model(s). It requires a test dataset.
}
