% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm_effort_offset.R
\name{effortOffset}
\alias{effortOffset}
\title{Function for generating a spatial thinned point process to reflect effort.}
\usage{
effortOffset(
  presences,
  window,
  thin.prob = TRUE,
  buffer = NULL,
  eps = 1e-09,
  ...
)
}
\arguments{
\item{presences}{A two column data.frame or matrix of presence locations. Typically we might used the presences for all species that fall in a particular taxa (say all reptiles).}

\item{window}{A spatial window at the same resolution and extent as the data is to be modelled.}

\item{thin.prob}{returned the plug in thinned probability for to use in the offset.}

\item{buffer}{A spatial distance (on the scale of the raster resolution) to buffer around each point. This will count the number of points per cell with the size of the point buffered to a radius of 'buffer' size.}

\item{eps}{A small value to use as the zero for the thinned probability.}
}
\description{
This function helps you set up a spatial thinning process (the probability that a cell has been surveyed) from known
presence-only occurrence data and a defined spatial window. The approach returns a thinning probability which is equad to: pi = 1 - exp(-lambda)
where lambda is the observed number of counts per cell. Clearly the cell resolution will effect the lambda values, so
make sure the window is the same spatial scale and resolution as the model you intend to fit.
}
\examples{
library(ppmData)
path <- system.file("extdata", package = "ppmData")
lst <- list.files(path=path,pattern='*.tif',full.names = TRUE)
win <- rast(lst[1])
presences <- pres <- as.matrix(snails[,1:2])
thpp <- effortOffset(pres,win)
}
\references{
Cressie, N. 1992. Statistics for spatial data. John Wiley & Sons..
}
\author{
Skipton Woolley
}
