% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm_fit.R
\name{ppmFit}
\alias{ppmFit}
\title{A function for fitting various ppms using a ppmData data object}
\usage{
ppmFit(
  species_formula = presence/weights ~ 1,
  bias_formula = NULL,
  ppmdata,
  family = c("poisson", "binomial"),
  link = c("default", "log", "logit", "cloglog"),
  method = c("lasso", "ridge"),
  control = list(),
  titbits = TRUE,
  ...
)
}
\arguments{
\item{species_formula}{The ppm model formula for the environmental/niche relationships. What drives the distribution of the species?}

\item{bias_formula}{Default is NULL. The idea will be to implement this as part of an integrated model.
Currently its just a nice way to keep track of which covariates are used in species or bias variables.
If you include this formula at the moment it will be merged into the species formula and used additively in a loglinear equation}

\item{ppmdata}{A ppmData data object}

\item{family}{Character What family to use, current options are "poisson" for a poisson point process and "binomial" of an infinitely Weighted Logistic Regression (IWLR).}

\item{link}{Character What link function to use? Uses "default", "log" for poisson or "logit" for binomial. A user can specific "clogclog" for a binomial model.}

\item{method}{A method to fit the a ppm. Default is 'lasso', the alternative option is 'ridge' for ridge regression.}

\item{control}{Options to pass to fitting functions.}

\item{\\dots}{Other things, not really used.}
}
\description{
This is a wrapper function which will fit a ppm model using a range of existing fitting tools.
}
\examples{
\dontrun{
library(ppmData)
library(terra)
path <- system.file("extdata", package = "ppmData")
lst <- list.files(path=path,pattern='*.tif',full.names = TRUE)
preds <- rast(lst)
presences <- subset(snails,SpeciesID \%in\% "Tasmaphena sinclairi")

ppmdata <- ppmData(npoints = 1000,presences=presences,
                  window = preds[[1]],
                  covariates = preds)

sp_form <- presence ~ poly(X,2) + poly(Y,2) + poly(max_temp_hottest_month,2)+
          poly(annual_mean_precip,2) + poly(annual_mean_temp,2) +
          poly(distance_from_main_roads,2)

ft.ppm1 <- ppmFit(species_formula = sp_form, ppmdata=ppmdata, method='lasso')
ft.ppm2 <- ppmFit(species_formula = sp_form, ppmdata=ppmdata, method='ridge')
ft.ppm3 <- ppmFit(species_formula = sp_form, ppmdata=ppmdata, method='lasso', family = 'binomial')
ft.ppm4 <- ppmFit(species_formula = sp_form, ppmdata=ppmdata, method='ridge', family = 'binomial', link="cloglog")
}
}
\references{
Berman, M. and Turner, T.R., 1992. Approximating point process likelihoods with GLIM. Journal of the Royal Statistical Society: Series C (Applied Statistics), 41(1), pp.31-38.

Warton, D.I. and Shepherd, L.C., 2010. Poisson point process models solve the" pseudo-absence problem" for presence-only data in ecology. The Annals of Applied Statistics, pp.1383-1402. \url{https://doi.org/10.1214/10-AOAS331}

Renner, I.W. and Warton, D.I., 2013. Equivalence of MAXENT and Poisson point process models for species distribution modeling in ecology. Biometrics, 69(1), pp.274-281.
#'@details Uses the Berman-Turner device to fit an approximate loglike for PPM using a weighted Poisson model.
}
\author{
Skipton Woolley
}
