% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm_cv.R
\name{pThin}
\alias{pThin}
\title{A function for p-thinning a point process}
\usage{
pThin(object, p = 0.25, seed = NULL, ...)
}
\arguments{
\item{object}{ppmData object that represents our species point process}

\item{p}{double The thinning probability (Bernoulli variable).}

\item{seed}{integer A seed if you want to set seed for randomisation}
}
\description{
We need to run a monte-carlo/kfold cross validation on the observed
point process. To do so requires the thinning of the observed pp. We can do
this via independent p-thinning. If the thinning is independent then the
resulting thinned point process can be used as another point process!
The current realization we have is the observed species locations. So if we
thin this based on a p-thinning and use the retained points as the validation
set and the thinned (removed) as the testing set we should be right.
}
\author{
Skipton Woolley
}
