% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm_predict.R
\name{transform}
\alias{transform}
\title{transform predictions from a ppmFit model.}
\usage{
transform(prediction, object, type = c("log", "logit", "cloglog"))
}
\arguments{
\item{prediction}{A prediction from a ppmFit model, can either be a terra "SpatRaster" or data.frame/matrix}

\item{object}{A fitted ppmFit model.}

\item{type}{What why of transform to do? The options are 'log', 'logit' and 'cloglog'. It assumes the input is the intensity prediction from the models.}
}
\description{
Sometimes we want to transform the expectation/intensity of a point process
to something that scales between zero and one. Typically we might do this with a
logistic or complementary log-log transform. One challenge with this approach is that we often
don't know the true prevalence of a species in the landscape. We can thus transform the
intensity to a 'probability of presence' and can use the log(Lambda) as an constant offset based
the expected count. Ideally you need presence-absence data to working out this prevalence, but
this is typically missing.
}
