% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{xyFromWindow}
\alias{xyFromWindow}
\title{generate lat/lon coordinates from raster/window}
\usage{
xyFromWindow(window, coord.name = c("X", "Y"), mask.na = FALSE)
}
\arguments{
\item{window}{rast A terra rast object}

\item{coord.name}{character The names to assign coordinated; default is "X" and "Y"}

\item{mask.na}{bool Do you want to mask the raster if NA's exist in the window?}
}
\description{
generate lat/lon coordinates from raster/window
}
\examples{
library(ppmData)
library(terra)
path <- system.file("extdata", package = "ppmData")
lst <- list.files(path=path,pattern='*.tif',full.names = TRUE)
covariates <- rast(lst)
xy.r <- xyFromWindow(covariates[[1]],mask.na=TRUE)
}
