% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm_cv.R
\name{cvfit}
\alias{cvfit}
\title{A function for doing a very basic cross validation fitting based on a ppmData data object}
\usage{
cvfit(
  species_formula,
  bias_formula,
  object,
  method,
  control,
  nsim,
  ncores,
  ...
)
}
\arguments{
\item{species_formula}{The ppm model formula.}

\item{bias_formula}{Default is NULL. The idea will be to implement this as part of an integrated model.
Currently its just a nice way to keep track of which covariates are used in species or bias variables.
If you include this formula at the moment it will be merged into the species formula.}

\item{object}{A ppmData object}

\item{method}{A method to fit the a ppm. Default is 'glm'. Others options are: 'gam','ppmlasso','lasso','ridge'}

\item{control}{Options to pass to fitting functions.}

\item{nsim}{Integer The number of cross validations to do (default is five).}

\item{ncores}{Integer The number of cores to used for parallel computing}

\item{\\dots}{Other things, not really used.}
}
\description{
This is a wrapper function which will fit a ppm model using a range of existing ppm fitting tools.
}
\examples{
\dontrun{
library(ppmData)
library(terra)
path <- system.file("extdata", package = "ppmData")
lst <- list.files(path=path,pattern='*.tif',full.names = TRUE)
preds <- rast(lst)
presences <- subset(snails,SpeciesID \%in\% "Tasmaphena sinclairi")
ppmdat <- ppmData(npoints = 10000,presences=presences, window = preds[[1]], covariates = preds)
species_formula <- presence ~ poly(X,2) + poly(Y,2) + poly(max_temp_hottest_month,2) + poly(annual_mean_precip,2) + poly(annual_mean_temp,2) + poly(distance_from_main_roads,2)
ft.ppm1 <- cvfit(species_formula = sp_form, object=ppmdat)
}
}
\author{
Skipton Woolley
}
